+++ magda-web-client/src/UI/QualityIndicator.js
@@ -1,8 +1,5 @@
 import React from "react";
-import { Link } from "react-router-dom";
-
-import "./QualityIndicator.css";
-import helpIcon from "../assets/help-24.svg";
+import Tooltip from "./Tooltip";
 
 function QualityIndicator(props) {
     let rating = Math.ceil(parseFloat(props.quality).toFixed(2) * 10 / 2) - 1;
@@ -39,13 +36,7 @@ function QualityIndicator(props) {
                     <span key={i} style={getBarColor(i)} />
                 ))}
             </span>&nbsp;
-            <div className="tooltip">
-                <img src={helpIcon} alt="Help Link" />
-                <span className="tooltiptext">
-                    Calculated using the{" "}
-                    <Link to="/page/dataset-quality">Open Data scale</Link>
-                </span>
-            </div>
+            <Tooltip/>
         </div>
     );
 }