+++ CHANGES.md
@@ -48,6 +48,7 @@
 *   Fixed an issue that users may see an error after page 4
 *   Fixed an issue that prevents users from searching current search text by clicking search button or pressing enter key
 *   keep distance between search box and the first story on mobile view
+*   Added quality tooltip to dataset details page
 
 ## 0.0.37